complete refactoring of grpc_plugin modules in a new, pythoning grpc_plugin_dll_class.py This should enable multiple desktop
