+++ _unittest/test_01_Design.py
@@ -305,6 +305,7 @@ class TestClass:
             "<class 'win32com.client.CDispatch'>",
             "<class 'PyDesktopPlugin.AedtObjWrapper'>",
             "<class 'pyaedt.generic.grpc_plugin.AedtObjWrapper'>",
+            "<class 'pyaedt.generic.grpc_plugin_dll_class.AedtObjWrapper'>",
         ]
 
     def test_28_get_pyaedt_app(self):
